requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
python-dotenv==1.2.1
python-multipart==0.0.20
#----------------------------------------
scikit-learn==1.7.2
xgboost==3.1.1
catboost==1.2.8
shap==0.50.0
lightgbm==4.6.0
#----------------------------------------
optuna==4.6.0
mlflow==3.8.1
#----------------------------------------
matplotlib==3.10.7
seaborn==0.13.2
PyYAML==6.0.3
